+++ android-app/app/src/main/java/ru/netology/nmedia/service/DateTimeService.kt
@@ -8,7 +8,7 @@ import java.util.Locale
 object DateTimeService {
     // Конвертирует и форматирует UnixTime в строку с датой и временем публикации
     fun formatUnixTime(unixDateTime: Long): String {
-        val instant = Instant.ofEpochSecond(unixDateTime.toLong())
+        val instant = Instant.ofEpochSecond(unixDateTime)
         val formatter = DateTimeFormatter.ofPattern("dd MMMM yyyy 'в' HH:mm", Locale("ru"))
             .withZone(ZoneId.systemDefault())
         return formatter.format(instant)
